# lib/db/src/schema/guildConfig.ts
import { pgTable, text, boolean, integer } from "drizzle-orm/pg-core";

export const guildConfigTable = pgTable("GuildConfig", {
  id: text("id").primaryKey(),
  guildId: text("guildId").notNull().unique(),

  ticketChannel: text("ticketChannel"),
  ticketCategory: text("ticketCategory"),
  ticketColor: text("ticketColor"),
  ticketBanner: text("ticketBanner"),
  ticketThumb: text("ticketThumb"),
  ticketFooter: text("ticketFooter"),
  ticketTitle: text("ticketTitle"),
  ticketText: text("ticketText"),
  ticketPingRole: text("ticketPingRole"),
  ticketPingUser: text("ticketPingUser"),
  ticketBtnLabel: text("ticketBtnLabel"),
  ticketBtnEmoji: text("ticketBtnEmoji"),
  ticketBtnStyle: text("ticketBtnStyle"),
  ticketOpenText: text("ticketOpenText"),
  ticketUseSeparator: boolean("ticketUseSeparator").notNull().default(false),
  ticketBannerPosition: text("ticketBannerPosition").notNull().default("top"),
  ticketOnlyBanner: boolean("ticketOnlyBanner").notNull().default(false),
  ticketUseMenu: boolean("ticketUseMenu").notNull().default(false),
  ticketQuestion1: text("ticketQuestion1"),
  ticketQuestion2: text("ticketQuestion2"),
  ticketQuestion3: text("ticketQuestion3"),

  instaChannel: text("instaChannel"),
  instaColor: text("instaColor"),
  instaEmoji: text("instaEmoji").notNull().default("💜"),
  instaHandle: text("instaHandle"),

  tellonymChannel: text("tellonymChannel"),
  tellonymColor: text("tellonymColor"),
  tellonymBanner: text("tellonymBanner"),
  tellonymThumb: text("tellonymThumb"),
  tellonymFooter: text("tellonymFooter"),
  tellonymTitle: text("tellonymTitle"),
  tellonymText: text("tellonymText"),
  tellonymBtnLabel: text("tellonymBtnLabel"),
  tellonymBtnEmoji: text("tellonymBtnEmoji"),

  lojaTitle: text("lojaTitle"),
  lojaText: text("lojaText"),
  lojaBanner: text("lojaBanner"),
  lojaThumb: text("lojaThumb"),
  lojaColor: text("lojaColor"),
  lojaConversao: text("lojaConversao"),
  lojaUseDivider: boolean("lojaUseDivider").notNull().default(false),
  shopEmojiComprar: text("shopEmojiComprar"),
  shopEmojiVitrine: text("shopEmojiVitrine"),
  shopEmojiConverter: text("shopEmojiConverter"),
  shopEmojiSaldo: text("shopEmojiSaldo"),
  shopEmojiGift: text("shopEmojiGift"),

  welcomeEnabled: boolean("welcomeEnabled").notNull().default(true),
  welcomeChannel: text("welcomeChannel"),
  welcomeColor: text("welcomeColor"),
  welcomeBanner: text("welcomeBanner"),
  welcomeThumb: text("welcomeThumb"),
  welcomeFooter: text("welcomeFooter"),
  welcomeTitle: text("welcomeTitle"),
  welcomeText: text("welcomeText"),
  welcomeRoles: text("welcomeRoles"),
  welcomeChannels: text("welcomeChannels"),
  welcomeUseDivider: boolean("welcomeUseDivider").notNull().default(false),
  welcomeBannerPosition: text("welcomeBannerPosition").notNull().default("top"),
  welcomeShowTitle: boolean("welcomeShowTitle").notNull().default(true),
  welcomeShowAvatar: boolean("welcomeShowAvatar").notNull().default(true),
  welcomeDeleteAfter: integer("welcomeDeleteAfter"),

  partnerEnabled: boolean("partnerEnabled").notNull().default(false),
  partnerChannel: text("partnerChannel"),
  partnerResponsibleRole: text("partnerResponsibleRole"),
  partnerPingRole: text("partnerPingRole"),
  partnerRole: text("partnerRole"),
  partnerNotifyDm: boolean("partnerNotifyDm").notNull().default(false),
  partnerMessage: text("partnerMessage"),
  partnerImage: text("partnerImage"),
  partnerThumbnail: text("partnerThumbnail"),
  partnerFooter: text("partnerFooter"),
  partnerColor: text("partnerColor"),
  partnerRemoveOnLeave: boolean("partnerRemoveOnLeave").notNull().default(false),

  botIconUrl: text("botIconUrl"),
  botBannerUrl: text("botBannerUrl"),
  botBio: text("botBio"),
  aiChannelId: text("aiChannelId"),

  ticketPanelMessageId: text("ticketPanelMessageId"),
  ticketPanelChannelId: text("ticketPanelChannelId"),
});

export type GuildConfig = typeof guildConfigTable.$inferSelect;
export type InsertGuildConfig = typeof guildConfigTable.$inferInsert;